removes util_test.go
